Refactor ignition helpers out of bootstrap for reuse
Functions like addStorageFiles() and addSystemdUnits() could be useful
for a new target (e.g. an all-in-one ignition).
