+++ pkg/asset/ignition/config.go
@@ -1,0 +1,217 @@
+package ignition
+
+import (
+	"bytes"
+	"io"
+	"io/ioutil"
+	"path"
+	"path/filepath"
+	"strings"
+	"text/template"
+
+	"github.com/coreos/ignition/config/util"
+	igntypes "github.com/coreos/ignition/config/v2_2/types"
+	"github.com/sirupsen/logrus"
+
+	"github.com/openshift/installer/data"
+	"github.com/openshift/installer/pkg/asset"
+)
+
+func AddStorageFiles(c *igntypes.Config, base string, uri string, templateData interface{}) (err error) {
+	file, err := data.Assets.Open(uri)
+	if err != nil {
+		return err
+	}
+	defer file.Close()
+
+	info, err := file.Stat()
+	if err != nil {
+		return err
+	}
+
+	if info.IsDir() {
+		children, err := file.Readdir(0)
+		if err != nil {
+			return err
+		}
+		if err = file.Close(); err != nil {
+			return err
+		}
+
+		for _, childInfo := range children {
+			name := childInfo.Name()
+			err = AddStorageFiles(c, path.Join(base, name), path.Join(uri, name), templateData)
+			if err != nil {
+				return err
+			}
+		}
+		return nil
+	}
+
+	name := info.Name()
+	_, data, err := readFile(name, file, templateData)
+	if err != nil {
+		return err
+	}
+
+	filename := path.Base(uri)
+	parentDir := path.Base(path.Dir(uri))
+
+	var mode int
+	appendToFile := false
+	if parentDir == "bin" || parentDir == "dispatcher.d" {
+		mode = 0555
+	} else if filename == "motd" {
+		mode = 0644
+		appendToFile = true
+	} else {
+		mode = 0600
+	}
+	ign := FileFromBytes(strings.TrimSuffix(base, ".template"), "root", mode, data)
+	ign.Append = appendToFile
+
+	// Replace files that already exist in the slice with ones added later, otherwise append them
+	c.Storage.Files = ReplaceOrAppend(c.Storage.Files, ign)
+
+	return nil
+}
+
+func AddSystemdUnits(c *igntypes.Config, uri string, templateData interface{}, enabled map[string]struct{}) (err error) {
+	directory, err := data.Assets.Open(uri)
+	if err != nil {
+		return err
+	}
+	defer directory.Close()
+
+	children, err := directory.Readdir(0)
+	if err != nil {
+		return err
+	}
+
+	for _, childInfo := range children {
+		dir := path.Join(uri, childInfo.Name())
+		file, err := data.Assets.Open(dir)
+		if err != nil {
+			return err
+		}
+		defer file.Close()
+
+		info, err := file.Stat()
+		if err != nil {
+			return err
+		}
+
+		if info.IsDir() {
+			if dir := info.Name(); !strings.HasSuffix(dir, ".d") {
+				logrus.Tracef("Ignoring internal asset directory %q while looking for systemd drop-ins", dir)
+				continue
+			}
+
+			children, err := file.Readdir(0)
+			if err != nil {
+				return err
+			}
+			if err = file.Close(); err != nil {
+				return err
+			}
+
+			dropins := []igntypes.SystemdDropin{}
+			for _, childInfo := range children {
+				file, err := data.Assets.Open(path.Join(dir, childInfo.Name()))
+				if err != nil {
+					return err
+				}
+				defer file.Close()
+
+				childName, contents, err := readFile(childInfo.Name(), file, templateData)
+				if err != nil {
+					return err
+				}
+
+				dropins = append(dropins, igntypes.SystemdDropin{
+					Name:     childName,
+					Contents: string(contents),
+				})
+			}
+
+			name := strings.TrimSuffix(childInfo.Name(), ".d")
+			unit := igntypes.Unit{
+				Name:    name,
+				Dropins: dropins,
+			}
+			if _, ok := enabled[name]; ok {
+				unit.Enabled = util.BoolToPtr(true)
+			}
+			c.Systemd.Units = append(c.Systemd.Units, unit)
+		} else {
+			name, contents, err := readFile(childInfo.Name(), file, templateData)
+			if err != nil {
+				return err
+			}
+
+			unit := igntypes.Unit{
+				Name:     name,
+				Contents: string(contents),
+			}
+			if _, ok := enabled[name]; ok {
+				unit.Enabled = util.BoolToPtr(true)
+			}
+			c.Systemd.Units = append(c.Systemd.Units, unit)
+		}
+	}
+
+	return nil
+}
+
+// Read data from the string reader, and, if the name ends with
+// '.template', strip that extension from the name and render the
+// template.
+func readFile(name string, reader io.Reader, templateData interface{}) (finalName string, data []byte, err error) {
+	data, err = ioutil.ReadAll(reader)
+	if err != nil {
+		return name, []byte{}, err
+	}
+
+	if filepath.Ext(name) == ".template" {
+		name = strings.TrimSuffix(name, ".template")
+		tmpl := template.New(name)
+		tmpl, err := tmpl.Parse(string(data))
+		if err != nil {
+			return name, data, err
+		}
+		stringData := applyTemplateData(tmpl, templateData)
+		data = []byte(stringData)
+	}
+
+	return name, data, nil
+}
+
+func applyTemplateData(template *template.Template, templateData interface{}) string {
+	buf := &bytes.Buffer{}
+	if err := template.Execute(buf, templateData); err != nil {
+		panic(err)
+	}
+	return buf.String()
+}
+
+func ReplaceOrAppend(files []igntypes.File, file igntypes.File) []igntypes.File {
+	for i, f := range files {
+		if f.Node.Path == file.Node.Path {
+			files[i] = file
+			return files
+		}
+	}
+	files = append(files, file)
+	return files
+}
+
+func AddParentFiles(c *igntypes.Config, dependencies asset.Parents, pathPrefix string, username string, mode int, assets []asset.WritableAsset) {
+	for _, asset := range assets {
+		dependencies.Get(asset)
+
+		// Replace files that already exist in the slice with ones added later, otherwise append them
+		for _, file := range FilesFromAsset(pathPrefix, username, mode, asset) {
+			c.Storage.Files = ReplaceOrAppend(c.Storage.Files, file)
+		}
+	}
+}
